requirements.txt
asgiref==3.11.1
diff-match-patch==20241021
Django==6.0.4
django-cors-headers==4.9.0
django-import-export==4.4.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.2.0
PyJWT==2.12.1
sqlparse==0.5.5
tablib==3.9.0
